refactor(user): 기존 UserRepository를 UserReaderImpl / UserWriterImpl로 분리
- UserRepository에서 읽기/쓰기 책임을 분리하여 CQRS 스타일 적용
- 읽기 전용 기능(findById, findByEmail, existsById 등)은 UserReaderImpl에서 처리
- 쓰기 전용 기능(saveUser, deleteUser)은 UserWriterImpl에서 처리
- 각각 UserReader, UserWriter 인터페이스를 구현하도록 구조 정리

고민한 점:
- 기존 UserRepository에 모든 메서드가 몰려 있어 책임 분리가 불명확하고 테스트 단위가 커졌음
- 조회 서비스에서 삭제나 저장 기능이 불필요하게 주입되는 문제가 있었음
- CQRS 원칙에 따라 의존성 주입 대상도 Reader / Writer로 명확히 구분함으로써
  단방향 의존성과 순환 참조 방지를 동시에 달성할 수 있었음
- 다만, 추후 기능이 겹치거나 Reader/Writer 간 중복이 발생할 경우
  Facade 혹은 Assembler 형태로 중간 조정 역할 도입을 고려해야 함
